fix : 충돌 해결
