+++ src/main/java/com/airbng/service/LockerServiceImpl.java
@@ -1,11 +1,17 @@
 package com.airbng.service;
 
+import com.airbng.common.exception.LockerException;
+import com.airbng.domain.base.ReservationState;
+import com.airbng.dto.LockerPreviewResult;
+import com.airbng.dto.LockerTop5Response;
 import com.airbng.dto.UserFindByIdResponse;
 import com.airbng.mappers.LockerMapper;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+
+import static com.airbng.common.response.status.BaseResponseStatus.NOT_FOUND_LOCKER;
 
 @Service
 @RequiredArgsConstructor
@@ -31,4 +37,14 @@ public class LockerServiceImpl implements LockerService{
                 .build();
     }
 
+    @Override
+    public LockerTop5Response findTop5Locker(){
+        List<LockerPreviewResult> popularLockers = lockerMapper.findTop5Lockers(ReservationState.CONFIRMED);
+
+        if(popularLockers.isEmpty()) throw new LockerException(NOT_FOUND_LOCKER);
+
+        return LockerTop5Response.builder()
+                .lockers(popularLockers)
+                .build();
+    }
 }
